fix: improved UX of file upload knowledge connectors
- Fixes after rebase
