+++ platform/backend/src/models/kb-document.ts
@@ -53,6 +53,7 @@ class KbDocumentModel {
         acl: schema.kbDocumentsTable.acl,
         metadata: schema.kbDocumentsTable.metadata,
         embeddingStatus: schema.kbDocumentsTable.embeddingStatus,
+        embeddingError: schema.kbDocumentsTable.embeddingError,
         chunkCount: schema.kbDocumentsTable.chunkCount,
         createdAt: schema.kbDocumentsTable.createdAt,
         updatedAt: schema.kbDocumentsTable.updatedAt,
@@ -110,6 +111,7 @@ class KbDocumentModel {
         acl: schema.kbDocumentsTable.acl,
         metadata: schema.kbDocumentsTable.metadata,
         embeddingStatus: schema.kbDocumentsTable.embeddingStatus,
+        embeddingError: schema.kbDocumentsTable.embeddingError,
         chunkCount: schema.kbDocumentsTable.chunkCount,
         createdAt: schema.kbDocumentsTable.createdAt,
         updatedAt: schema.kbDocumentsTable.updatedAt,
@@ -293,6 +295,7 @@ class KbDocumentModel {
         acl: schema.kbDocumentsTable.acl,
         metadata: schema.kbDocumentsTable.metadata,
         embeddingStatus: schema.kbDocumentsTable.embeddingStatus,
+        embeddingError: schema.kbDocumentsTable.embeddingError,
         chunkCount: schema.kbDocumentsTable.chunkCount,
         createdAt: schema.kbDocumentsTable.createdAt,
         updatedAt: schema.kbDocumentsTable.updatedAt,
