+++ platform/backend/src/routes/knowledge-base.ts
@@ -1656,6 +1656,7 @@ async function enrichKnowledgeFiles(params: {
     processingStatus: file.processingStatus,
     processingError: file.processingError ?? null,
     embeddingStatus: docByFileId.get(file.id)?.embeddingStatus ?? "pending",
+    embeddingError: docByFileId.get(file.id)?.embeddingError ?? null,
     assignedAgents: (agentIdsByConnector.get(file.connectorId) ?? []).flatMap(
       (id) => {
         const agent = agentById.get(id);
