+++ platform/shared/hey-api/clients/api/types.gen.ts
@@ -33423,6 +33423,7 @@ export type GetConnectorDocumentsResponses = {
                 [key: string]: unknown;
             } | null;
             embeddingStatus: 'pending' | 'processing' | 'completed' | 'failed';
+            embeddingError: string | null;
             chunkCount: number;
             createdAt: string;
             updatedAt: string;
@@ -33620,6 +33621,7 @@ export type GetConnectorDocumentResponses = {
             [key: string]: unknown;
         } | null;
         embeddingStatus: 'pending' | 'processing' | 'completed' | 'failed';
+        embeddingError: string | null;
         chunkCount: number;
         createdAt: string;
         updatedAt: string;
@@ -34845,6 +34847,7 @@ export type GetKnowledgeFileResponses = {
         processingStatus: 'pending' | 'processing' | 'completed' | 'failed';
         processingError: string | null;
         embeddingStatus: 'pending' | 'processing' | 'completed' | 'failed';
+        embeddingError: 'api_unauthorized' | 'api_permission_denied' | 'api_bad_request' | 'api_not_found' | 'api_conflict' | 'api_unprocessable_entity' | 'api_rate_limit' | 'api_generic_error' | 'context_length_exceeded' | 'length_mismatch' | 'dimensions_mismatch' | 'unknown';
         assignedAgents: Array<{
             id: string;
             name: string;
